Refactoring ReportStreamHeader into separate elements
