+++ frontend-react/src/pages/daily/Table/TableReports.tsx
@@ -26,12 +26,6 @@ function TableReports({ sortBy }: { sortBy?: string }) {
                 <h2>Test results</h2>
                 {
                     /* Button group only shows when there is more than a single sender. */
-
-                    /*  
-                        TODO: This may need to be remedied not to show "sendingOrg" as a unique sender
-                        as most data input thus far does not have a proper sendingOrg property.
-                        >> 09/09/2021 (Kevin Haube)
-                    */
                     receiverSVCs.length > 1 ?
                         <TableButtonGroup senders={receiverSVCs} chosenCallback={handleCallback} />
                         :
